+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 Audiograph adheres to [Semantic Versioning](http://semver.org/).
 
 ## Next Release:
+
+## 0.4.0 (29.03.2020):
 - Refactored scaling timestamps in DataProcessor.
 - Fixed potential heap overflow.
 - Fixed a bug where completion block was not called on main queue every time.